LayoutTests/inspector/canvas/recording-offscreen-bitmaprenderer-full-expected.txt

Test that CanvasManager is able to record actions made to offscreen BitmapRenderer canvas contexts.


== Running test suite: Canvas.recordingOffscreenBitmapRenderer
-- Running test case: Canvas.recordingOffscreenBitmapRenderer.multipleFrames
initialState:
  attributes:
    width: 2
    height: 2
  parameters:
    0: {"alpha":true}
  content: <filtered>
frames:
  0: (duration)
    0: transferFromImageBitmap([object ImageBitmap])
      swizzleTypes: [ImageBitmap]
      trace:
        0: transferFromImageBitmap
        1: (anonymous function)
        2: executeFrameFunction
        3: (anonymous function)
      snapshot: <filtered>
  1: (duration)
    0: width
      trace:
        0: (anonymous function)
        1: executeFrameFunction
    1: width = 2
      swizzleTypes: [Number]
      trace:
        0: (anonymous function)
        1: executeFrameFunction
  2: (duration)
    0: height
      trace:
        0: (anonymous function)
        1: executeFrameFunction
    1: height = 2
      swizzleTypes: [Number]
      trace:
        0: (anonymous function)
        1: executeFrameFunction
